feat(ARC-2228): extract iiif as separate component (wip)
https://meemoo.atlassian.net/browse/ARC-2228
