+++ src/modules/ie-objects/ObjectDetailPage.helpers.tsx
