+++ src/modules/iiif-viewer/IiifViewer.module.scss
@@ -8,20 +8,6 @@ $metadata-collapsed-width: 36vw;
 	width: 100%;
 	height: 100%;
 
-	:global(.mirador-window-top-bar) {
-		display: none;
-	}
-
-	:global(.mirador-canvas-nav) {
-		display: none;
-	}
-
-	:global(.mirador-window),
-	:global(.MuiDrawer-docked),
-	:global(.mirador-companion-area-left) {
-		z-index: 0 !important;
-	}
-
 	:global(.openseadragon-container) {
 		background-color: #222 !important;
 	}
@@ -118,28 +104,4 @@ $metadata-collapsed-width: 36vw;
 			outline: none;
 		}
 	}
-}
-
-.p-object-detail__metadata__ocr {
-	width: $metadata-collapsed-width;
-	padding: 6.4rem 3.2rem;
-	overflow-y: auto;
-
-	> :global(.p-object-detail__metadata__ocr__word) {
-		display: inline-block;
-		padding: 0.5rem 0.25rem;
-		cursor: pointer;
-
-		&:hover {
-			background-color: #00857d;
-			color: white;
-		}
-	}
-}
-
-.p-object-detail__metadata__ocr__close {
-	display: flex;
-	flex-direction: row;
-	justify-content: flex-end;
-	margin-bottom: 3.2rem;
 }
