+++ src/pages/zoeken/[slug]/[ie]/[name]/index.tsx
@@ -2,11 +2,11 @@ import { NextPage } from 'next';
 import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next/types';
 import React from 'react';
 
+import { ObjectDetailPage } from '@ie-objects/ObjectDetailPage';
+import { prefetchDetailPageQueries } from '@ie-objects/ObjectDetailPage.helpers';
 import { getIeObjectInfo } from '@ie-objects/hooks/get-ie-objects-info';
 import { IeObjectsService } from '@ie-objects/services';
 import { SeoInfo } from '@ie-objects/services/ie-objects/ie-objects.service.types';
-import { ObjectDetailPage } from '@search/ObjectDetailPage';
-import { prefetchDetailPageQueries } from '@search/ObjectDetailPage.helpers';
 import { getDefaultStaticProps } from '@shared/helpers/get-default-server-side-props';
 import { DefaultSeoInfo } from '@shared/types/seo';
